Fix using char for the return type of fgets(3)
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@33018 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/os2/rs232.c
@@ -128,7 +128,7 @@ int rs232_getc(int fi, BYTE *b)
     if (fd[fi] == NULL) {
         return -1;
     }
-    *b = fgetc(fd[fi]);
+    *b = (BYTE)fgetc(fd[fi]);
     return 0;
 }
